setUp server.js, setUp routs
