+++ public/assets/js/index.js
@@ -1,17 +1,3 @@
-const express = require('express');
-const path = require('path');
-const notesData = require('./db.json');
-const app = express();
-
-const PORT = 3001;
-
-// Middleware:
-// Middleware to use folder Public with static data
-app.use(express.static('public'));
-// Middleware to work with user's request using 'req.body' in a JSON format
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
 let noteForm;
 let noteTitle;
 let noteText;
@@ -42,28 +28,16 @@ const hide = (elem) => {
 // activeNote is used to keep track of the note in the textarea
 let activeNote = {};
 
-//routes
-
-// sourse: UofU bootcamp, module 11, activity 05
-// a get URL 'localhost:3001/notes' request from user will bring/display to user the notes.html file 
-app.get('/notes', (req, res) =>
-  res.sendFile(path.join(__dirname, 'public/notes.html'))
-);
-// a get URL 'localhost:3001/*' request from user will bring/display to user the index.html file. '*' - stands for any user input
-app.get('*', (req, res) =>
-  res.sendFile(path.join(__dirname, 'public/index.html'))
-);
-
-// ??? do I need this routh?
-// a get URL 'localhost:3001/api/notes' request from user will bring/display to user the db.json file with all saved notes as JSON.
-app.get('/api/notes', (req, res) => {
-  res.json(notesData)
-});
-
-
-
 const getNotes = () =>
   fetch('/api/notes', {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json'
+    }
+  });
+
+const getIdNote = (id) =>
+  fetch(`/api/notes/${id}`, {
     method: 'GET',
     headers: {
       'Content-Type': 'application/json'
